// SGUEES-SPA/src/app/pages/SelectionHiring/sc-tipo-contratacion/models/sc-tipo-contratacion.ts
export interface ScTipoContratacion {
	CORR_EMPRESA: number;
	CORR_TIPO_CONTRATACION: number;
	NOMBRE_TIPO_CONTRATACION: string;
	ES_PERMANENTE: boolean;
	AREA_APLICADA: string;
	ACTIVO: boolean;
	USUARIO_CREA: string;
	ESTACION_CREA: string;
	FECHA_CREA: Date;
	USUARIO_ACTU: string;
	ESTACION_ACTU: string;
	FECHA_ACTU: Date;
}